Move types to endpoint/types.ts, address PR comments
blah 2
