+++ x-pack/plugins/endpoint/public/applications/endpoint/store/alerts/action.ts
@@ -4,12 +4,11 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { AlertData } from '../../../../../common/types';
+import { AlertData, Immutable } from '../../../../../common/types';
 
-interface ServerReturnedAlertsData {
-  readonly type: 'serverReturnedAlertsData';
-
-  readonly payload: AlertData[];
-}
+type ServerReturnedAlertsData = Immutable<{
+  type: 'serverReturnedAlertsData';
+  payload: AlertData[];
+}>;
 
 export type AlertAction = ServerReturnedAlertsData;
