+++ x-pack/plugins/endpoint/public/applications/endpoint/store/alerts/index.ts
@@ -6,4 +6,3 @@
 
 export { alertListReducer } from './reducer';
 export { AlertAction } from './action';
-export * from './types';
