+++ x-pack/plugins/endpoint/public/applications/endpoint/store/alerts/middleware.ts
@@ -4,25 +4,16 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { Dispatch, MiddlewareAPI } from 'redux';
-import { CoreStart } from 'kibana/public';
-import { AlertData } from '../../../../../common/types';
-import { GlobalState } from '../reducer';
+import { AlertData, ImmutableArray } from '../../../../../common/types';
 import { AppAction } from '../action';
-
-// TODO, move this somewhere
-type MiddlewareFactory = (
-  coreStart: CoreStart
-) => (
-  api: MiddlewareAPI<Dispatch<AppAction>, GlobalState>
-) => (next: Dispatch<AppAction>) => (action: AppAction) => unknown;
+import { MiddlewareFactory } from '../../types';
 
 export const alertMiddlewareFactory: MiddlewareFactory = coreStart => {
-  return store => next => async (action: AppAction) => {
+  return api => next => async (action: AppAction) => {
     next(action);
     if (action.type === 'userNavigatedToPage' && action.payload === 'alertsPage') {
-      const response: AlertData[] = await coreStart.http.get('/api/endpoint/alerts');
-      store.dispatch({ type: 'serverReturnedAlertsData', payload: response });
+      const response: ImmutableArray<AlertData> = await coreStart.http.get('/api/endpoint/alerts');
+      api.dispatch({ type: 'serverReturnedAlertsData', payload: response });
     }
   };
 };
