+++ x-pack/plugins/endpoint/public/applications/endpoint/store/alerts/selectors.ts
@@ -4,6 +4,6 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { AlertListState } from './types';
+import { AlertListState } from '../../types';
 
 export const alertListData = (state: AlertListState) => state.alerts;
