--- x-pack/plugins/endpoint/public/applications/endpoint/store/alerts/types.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import { Immutable, AlertData } from '../../../../../common/types';
-
-export type AlertListState = Immutable<{
-  alerts: AlertData[];
-}>;
